Reapply "More strict typed usage of typescript (#319)"
This reverts commit c4b523e4bc0dad2a8094ff0589576e70cfe1670c.

+++ src/main/content/maps/map-image.ts
@@ -14,6 +14,9 @@ const promises = new Map<string, Promise<ArrayBuffer>>();
 
 worker.on("message", ({ imageSource, arrayBuffer }) => {
     const promiseHandles = jobs.get(imageSource);
+
+    if (!promiseHandles) throw new Error("failed to access image source promise handlers");
+
     promiseHandles.resolve(arrayBuffer);
     jobs.delete(imageSource);
     promises.delete(imageSource);
@@ -23,7 +26,7 @@ worker.on("message", ({ imageSource, arrayBuffer }) => {
 //     if (code !== 0) reject(new Error(`map-image-worker stopped with exit code ${code}`));
 // });
 
-export function fetchMapImages(imageSource: string): Promise<ArrayBuffer> {
+export function fetchMapImages(imageSource: string): Promise<ArrayBuffer> | undefined {
     if (promises.has(imageSource)) return promises.get(imageSource);
     const promise = new Promise<ArrayBuffer>((resolve, reject) => {
         jobs.set(imageSource, { resolve, reject });
